pyproject.toml
[tool.poetry]
name = "musicrecommender"
version = "0.1.0"
description = ""
authors = ["jakem <mazurkiewiczj@duq.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
pandas = "^2.2.2"
streamlit = "^1.37.0"
sentence-transformers = "^3.0.1"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
